Split code related to security headers into another module
